refactor: rename initialSection file
