+++ frontend/src/components/InitialPage/InitialSection/InitialSection.tsx
